Separate all redis stuff
