+++ cacheops/redis.py
@@ -1,0 +1,63 @@
+from __future__ import absolute_import
+import warnings
+
+from funcy import decorator, identity, memoize
+import redis
+from django.core.exceptions import ImproperlyConfigured
+
+from .conf import CACHEOPS_REDIS, CACHEOPS_DEGRADE_ON_FAILURE
+
+
+if CACHEOPS_DEGRADE_ON_FAILURE:
+    @decorator
+    def handle_connection_failure(call):
+        try:
+            return call()
+        except redis.ConnectionError as e:
+            warnings.warn("The cacheops cache is unreachable! Error: %s" % e, RuntimeWarning)
+        except redis.TimeoutError as e:
+            warnings.warn("The cacheops cache timed out! Error: %s" % e, RuntimeWarning)
+else:
+    handle_connection_failure = identity
+
+
+class SafeRedis(redis.StrictRedis):
+    get = handle_connection_failure(redis.StrictRedis.get)
+
+
+class LazyRedis(object):
+    def _setup(self):
+        if not CACHEOPS_REDIS:
+            raise ImproperlyConfigured('You must specify CACHEOPS_REDIS setting to use cacheops')
+
+        client = (SafeRedis if CACHEOPS_DEGRADE_ON_FAILURE else redis.StrictRedis)(**CACHEOPS_REDIS)
+
+        object.__setattr__(self, '__class__', client.__class__)
+        object.__setattr__(self, '__dict__', client.__dict__)
+
+    def __getattr__(self, name):
+        self._setup()
+        return getattr(self, name)
+
+    def __setattr__(self, name, value):
+        self._setup()
+        return setattr(self, name, value)
+
+redis_client = LazyRedis()
+
+
+### Lua script loader
+
+import re
+import os.path
+
+STRIP_RE = re.compile(r'TOSTRIP.*/TOSTRIP', re.S)
+
+@memoize
+def load_script(name, strip=False):
+    filename = os.path.join(os.path.dirname(__file__), 'lua/%s.lua' % name)
+    with open(filename) as f:
+        code = f.read()
+    if strip:
+        code = STRIP_RE.sub('', code)
+    return redis_client.register_script(code)
